Refactor tasks
